websocket/requirements.txt
fastapi
pymysql
sqlalchemy
cryptography
strawberry-graphql[fastapi]
uvicorn[standard]
aiokafka
pandas
